Improve performance of the mach port coordinator by reusing repeated event instead of resolving it again from the CGEvent Reduce casting between Int and Int64 (CGEventSignature now uses Int64)

+++ App/Sources/Core/MachPort/ShortcutResolver.swift
@@ -161,7 +161,7 @@ final class ShortcutResolver {
               flags.insert(.maskSecondaryFn)
             }
 
-            let eventSignature = CGEventSignature(keyCode, flags)
+            let eventSignature = CGEventSignature(Int64(keyCode), flags)
 
             if Self.debug, workflow.name.contains("**") {
               print(workflow.name, eventSignature.id)
